Rename AreWeGreenClient.js to AreWeGreenDataClient.js
