+++ src/main/reactjs/data/AreWeGreenDataClient.js
